+++ packages/domain/src/index.ts
@@ -5,5 +5,6 @@
 
 export * from './enums.js';
 export * from './identity.js';
+export * from './ids.js';
 export * from './schemas.js';
 export * from './types.js';